Improvements 01, Api in diferents files
Moving the files of api to the "api" directory
Create a new files into "api/src" called "index.js"
into "routes" (new directory) create two new files: departments.js and employee.js

#20211012-improvement01
