+++ api/src/database.js
@@ -1,0 +1,18 @@
+const mysql = require('mysql');
+const mysqlConnection = mysql.createConnection({
+  host: 'localhost',
+  user: 'root',
+  password: 'cascade',
+  database: 'mytestdb'
+});
+
+mysqlConnection.connect(function(err){
+  if(err) {
+    console.log(err);
+    return;
+  } else {
+    console.log('DB is connected');
+  }
+})
+
+module.exports = mysqlConnection;
